+++ spec/support/framework_status.rb
@@ -1,4 +1,4 @@
-%w[RM6238 RM6240 RM6309 RM6360 RM6378].each do |framework_id|
+%w[RM6309 RM6360 RM6374 RM6376 RM6378].each do |framework_id|
   RSpec.shared_context "and #{framework_id} is live in the future" do
     before { Framework.find(framework_id).update(live_at: 1.day.from_now) }
   end
@@ -12,7 +12,7 @@
   end
 end
 
-%w[RM6187 RM3830 RM6232].each do |framework_id|
+%w[RM6187 RM6238 RM3830 RM6232 RM6240].each do |framework_id|
   RSpec.shared_context "and #{framework_id} is live" do
     before { Framework.find(framework_id).update(expires_at: 1.day.from_now) }
   end